Do not deallocate the shared data in the interrupt context.
When the creator of a thread calls thread::detach(), the threaded function has to deallocate the
shared thread state. The deallocation should not be done in an interrupt context because the
allocator might want to lock a mutex. If it does so, a hard fault exception is triggered.

The solution is to deallocate the shared data before the threaded function cancels itself in an
interrupt context.